FCT: chargement durant tout les requetes

+++ DixSurDixGeoffrion/app/src/main/java/com/example/dixsurdixgeoffrion/Services/ServiceEpicerie.java
@@ -1,9 +1,11 @@
 package com.example.dixsurdixgeoffrion.Services;
-import android.app.Dialog;
 import android.net.Uri;
+import android.view.View;
+import android.widget.ProgressBar;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.dixsurdixgeoffrion.ListeDepicerie.AjoutAutoAdapter;
 import com.example.dixsurdixgeoffrion.ListeDepicerie.MainListeDepicerie;
@@ -20,18 +22,20 @@ import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
 import com.google.firebase.storage.UploadTask;
 
-import java.time.Instant;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 public class ServiceEpicerie {
 
-    private List<Aliment> alimentList = new ArrayList<>(); //Cette liste sert à envoyer une liste d'aliments dans la liste d'epicerie principal
+    public List<Aliment> alimentList = new ArrayList<>(); //Cette liste sert à envoyer une liste d'aliments dans la liste d'epicerie principal
     public List<Aliment> alimentListAuto = new ArrayList<>(); //Cette liste sert à ajouter des aliments automatiques selectionnés
+
+    public List<AlimentAuto> alimentAutoList = new ArrayList<>();
     private DatabaseReference _rootDataref;
     private StorageReference rootStorage;
     private MainListeDepicerie context;
+
+    public DialogService dialogService;
 
 
     public ServiceEpicerie(MainListeDepicerie current_context){
@@ -41,9 +45,11 @@ public class ServiceEpicerie {
     }
     public void GetListAliment(){
 
-        _rootDataref.child("AlimentsEpicerie").addValueEventListener(new ValueEventListener() {
+        dialogService.showDialogLoadingWaiting();
+
+        _rootDataref.child("AlimentsEpicerie").get().addOnSuccessListener(new OnSuccessListener<DataSnapshot>() {
             @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
+            public void onSuccess(DataSnapshot snapshot) {
                 if (snapshot.exists()){
                     Iterable<DataSnapshot> list = snapshot.getChildren();
 
@@ -51,32 +57,36 @@ public class ServiceEpicerie {
                         alimentList.add(s.getValue(Aliment.class));
                     }
                 }
-
+                //dialogService.dismissDialogLoadingWaiting();
                 context.remplirRecycler(alimentList);
-                alimentList.clear();
             }
-
+        }).addOnFailureListener(new OnFailureListener() {
             @Override
-            public void onCancelled(@NonNull DatabaseError error) {
+            public void onFailure(@NonNull Exception e) {
                 Toast.makeText(context,"Problème inattendue",Toast.LENGTH_LONG).show();
+                dialogService.dismissDialogLoadingWaiting();
             }
         });
     }
-    public void GetListAutoAliment(AjoutAutoAdapter ajoutAutoAdapter) {
+    public void GetListAutoAliment(AjoutAutoAdapter ajoutAutoAdapter, RecyclerView recyclerView, ProgressBar progressBar) {
         _rootDataref.child("Aliments Automatiques").addValueEventListener(new ValueEventListener() {
             @Override
             public void onDataChange(@NonNull DataSnapshot snapshot) {
                 if (snapshot.exists()){
                     Iterable<DataSnapshot> list = snapshot.getChildren();
                     for (DataSnapshot s : list){
-
                         AlimentAuto alimentAuto = s.getValue(AlimentAuto.class);
                         if (!alimentAuto.isUsed()){
                             ajoutAutoAdapter.listAliment.add(alimentAuto);
                         }
                     }
-
                     ajoutAutoAdapter.notifyDataSetChanged();
+                    recyclerView.setVisibility(View.VISIBLE);
+                    progressBar.setVisibility(View.GONE);
+                }else{
+                    recyclerView.setVisibility(View.VISIBLE);
+                    progressBar.setVisibility(View.GONE);
+                    Toast.makeText(context,"Erreur d'affiche des aliments auto", Toast.LENGTH_LONG).show();
                 }
             }
 
@@ -87,8 +97,6 @@ public class ServiceEpicerie {
         });
     }
     public void AjouterAliment(Aliment aliment, Uri imageUri){
-
-       // rootStorage = FirebaseStorage.getInstance().getReference().child("AlimentImages");
 
         //Création de la clé qui sera la même pour l'image et l'aliment
         String key = _rootDataref.push().getKey();
@@ -161,34 +169,36 @@ public class ServiceEpicerie {
         _rootDataref.child("AlimentsEpicerie/"+ aliment.alimentKey).setValue(aliment);
     }
     public void UpdateUsedAutoAliment(AlimentAuto alimentAuto){
-        _rootDataref.child("Aliments Automatiques/" + alimentAuto.alimentKey).setValue(alimentAuto);
+        _rootDataref.child("Aliments Automatiques").child(alimentAuto.alimentKey).setValue(alimentAuto);
     }
     public void SupprimerAliment(Aliment aliment){
-
-
         _rootDataref.child("AlimentsEpicerie/"+ aliment.alimentKey).removeValue().addOnSuccessListener(new OnSuccessListener<Void>() {
             @Override
             public void onSuccess(Void unused) {
                 if (aliment.alimentauto){
-                    _rootDataref.child("Aliments Automatiques").addValueEventListener(new ValueEventListener() {
+                    _rootDataref.child("Aliments Automatiques").get().addOnSuccessListener(new OnSuccessListener<DataSnapshot>() {
                         @Override
-                        public void onDataChange(@NonNull DataSnapshot snapshot) {
+                        public void onSuccess(DataSnapshot snapshot) {
                             Iterable<DataSnapshot> list = snapshot.getChildren();
+                            AlimentAuto alimentAuto = null;
                             for (DataSnapshot s : list){
                                 if (s.getValue(AlimentAuto.class).nom.equals(aliment.nom)){
-                                    AlimentAuto alimentAuto = s.getValue(AlimentAuto.class);
+                                    alimentAuto = s.getValue(AlimentAuto.class);
                                     alimentAuto.used = false;
                                     alimentAuto.quantite = 0;
                                     alimentAuto.validerAchat = false;
-                                    _rootDataref.child("Aliments Automatiques/" + alimentAuto.alimentKey).setValue(alimentAuto);
-                                    GetListAliment();
                                     break;
                                 }
                             }
-                        }
+                            if (alimentAuto != null){
+                                _rootDataref.child("Aliments Automatiques/" + alimentAuto.alimentKey).setValue(alimentAuto);
+                                GetListAliment();
+                            }
 
+                        }
+                    }).addOnFailureListener(new OnFailureListener() {
                         @Override
-                        public void onCancelled(@NonNull DatabaseError error) {
+                        public void onFailure(@NonNull Exception e) {
 
                         }
                     });
@@ -207,16 +217,45 @@ public class ServiceEpicerie {
         });
     }
     public void SupprimerToutLesAliments(List<Aliment> aliments){
+        _rootDataref.child("Aliments Automatiques").addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot snapshot) {
 
+                //Comme la liste aliments provient de l'adapter, on supprime le premier element qui est null.
+                if (aliments.size() > 0){
+                    aliments.remove(0);
+                }
 
+                Iterable<DataSnapshot> list = snapshot.getChildren();
+                List<AlimentAuto> alimentAutos = new ArrayList<>();
+                for (DataSnapshot s : list){
+                    alimentAutos.add(s.getValue(AlimentAuto.class));
+                }
 
-        //Comme la liste aliments provient de l'adapter, on supprime le premier element qui est null.
-        aliments.remove(0);
-        //Suppresion dans le storage
-        for (Aliment aliment : aliments) {
-            rootStorage.child("AlimentsManuels").child(aliment.alimentKey + ".jpg").delete();
-        }
-        _rootDataref.child("AlimentsEpicerie").removeValue();
+                for(Aliment aliment : aliments){
+                    _rootDataref.child("AlimentsEpicerie").child(aliment.alimentKey).removeValue();
+                    if (aliment.alimentauto){
+                        for (AlimentAuto alimentAuto : alimentAutos){
+                            if (alimentAuto.nom.equals(aliment.nom)){
+                                alimentAuto.used = false;
+                                alimentAuto.quantite = 0;
+                                alimentAuto.validerAchat = false;
+                                _rootDataref.child("Aliments Automatiques/" + alimentAuto.alimentKey).setValue(alimentAuto);
+                            }
+                        }
+                    }else{
+                        rootStorage.child("AlimentsManuels").child(aliment.alimentKey + ".jpg").delete();
+                    }
+                }
+
+                GetListAliment();
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError error) {
+
+            }
+        });
         
     }
